materijali/pred-10/pyproject.toml
[tool.poetry]
name = "pdaj-distributed"
version = "0.1.0"
description = "distributed demo for pdaj predavanja"
authors = ["Petar Marić <petar.maric@gmail.com>"]
license = "BSD"

[tool.poetry.dependencies]
python = "^3.6"
celery = {version = "~=5.0.0", extras = ["redis"]}
python-decouple = "~=3.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
